+++ reactivestate/src/main/java/com/ensody/reactivestate/DerivedLiveData.kt
@@ -5,13 +5,15 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import kotlinx.coroutines.CoroutineScope
 
-// TODO: Make this evaluate lazily and non-null
 class DerivedLiveData<T>(scope: CoroutineScope, private val observer: AutoRunCallback<T>) :
     LiveData<T>(null) {
 
     init {
         scope.autoRun { value = observer() }
     }
+
+    @Suppress("UNCHECKED_CAST")
+    override fun getValue(): T = super.getValue() as T
 }
 
 fun <T> CoroutineScope.derived(observer: AutoRunCallback<T>) =